ISSUE #5544 - refactor
